fix: Address clippy warnings across codebase
- Replace .min().max() patterns with .clamp()
- Fix operator precedence issues with bit shifts
- Add type aliases for complex biquad coefficient types
- Use .flatten() for iterator patterns
- Remove unnecessary borrows and casts
- Replace manual range checks with RangeInclusive::contains()
- Fix nested if let patterns to collapse into single pattern

+++ src/hardware/hid.rs
@@ -139,7 +139,7 @@ pub fn pull_peq_internal(
         }
         filter_responses[i as usize] = response;
 
-        delay_ms(cfg.inter_filter_ms as u64);
+        delay_ms(cfg.inter_filter_ms);
     }
 
     if had_mismatch {
